--- backend/api/canvasApiClient.js
@@ -1,265 +1,0 @@
-const Canvas = require("@kth/canvas-api");
-const FormData = require("formdata-node").default;
-const fs = require("fs");
-const got = require("got");
-const log = require("skog");
-const { getAktivitetstillfalle } = require("./ladokApiClient");
-
-const canvas = new Canvas(
-  process.env.CANVAS_API_URL,
-  process.env.CANVAS_API_ADMIN_TOKEN
-);
-
-/** Get data from one canvas course */
-async function getCourse(courseId) {
-  const { body } = await canvas.get(`courses/${courseId}`);
-
-  return body;
-}
-
-/** Creates a "good-looking" homepage in Canvas */
-async function createHomepage(courseId) {
-  await canvas.requestUrl(`courses/${courseId}/front_page`, "PUT", {
-    wiki_page: {
-      // To make this page, use the Rich Content Editor in Canvas (https://kth.test.instructure.com/courses/30347/pages/welcome-to-the-exam/edit)
-      // Then copy the HTML code:
-      body: `<p>Welcome to the Canvas page for the exam results</p>
-      <p>This course will be used to grade your exams digitally. This means that, after your exams are scanned, they will be uploaded in this Canvas course</p>
-      <p>&nbsp;</p>
-      <p>Once your exam has been graded, you will be able to see the grades and teachers feedback under "Grades".</p>`,
-    },
-  });
-  return canvas.requestUrl(`courses/${courseId}`, "PUT", {
-    course: {
-      default_view: "wiki",
-    },
-  });
-}
-
-/** Publish a course */
-async function publishCourse(courseId) {
-  return canvas.requestUrl(`courses/${courseId}`, "PUT", {
-    course: {
-      event: "offer",
-    },
-  });
-}
-
-/** Get the Ladok UID of the examination linked with a canvas course */
-async function getExaminationLadokId(courseId) {
-  const sections = await canvas.list(`courses/${courseId}/sections`).toArray();
-
-  // For SIS IDs with format "AKT.<ladok id>.<suffix>", take the "<ladok id>"
-  const REGEX = /^AKT\.([\w-]+)/;
-  const sisIds = sections
-    .map((section) => section.sis_section_id?.match(REGEX)?.[1])
-    .filter((sisId) => sisId /* Filter out null and undefined */);
-
-  // Deduplicate IDs (there are usually one "funka" and one "non-funka" with
-  // the same Ladok ID)
-  const uniqueIds = Array.from(new Set(sisIds));
-
-  // Right now we are not supporting rooms with more than one examination
-  if (uniqueIds.length !== 1) {
-    throw new Error(
-      `Course ${courseId} not supported: it is connected to ${uniqueIds.length} different Ladok Ids`
-    );
-  } else {
-    return uniqueIds[0];
-  }
-}
-
-async function getValidAssignment(courseId, ladokId) {
-  const assignments = await canvas
-    .list(`courses/${courseId}/assignments`)
-    .toArray();
-
-  // TODO: Filter more strictly?
-  return (
-    assignments.find((a) => a.integration_data?.ladokId === ladokId) ?? null
-  );
-}
-
-async function createAssignment(courseId, ladokId) {
-  const examination = await getAktivitetstillfalle(ladokId);
-
-  return canvas
-    .requestUrl(`courses/${courseId}/assignments`, "POST", {
-      assignment: {
-        name: "Scanned exams",
-        description:
-          'This is an assignment created automatically by importing scanned exams to Canvas. The grade posting policy is set to "Manual" which makes it possible to grade all submissions before publishing the student feedback for all students at once.',
-        submission_types: ["online_upload"],
-        allowed_extensions: ["pdf"],
-        // TODO: save only the "Ladok UID" because `examination.courseCode` and
-        //       `examination.examCode` can be more than one
-        // TODO: add more data to be able to filter out better?
-        integration_data: {
-          ladokId,
-        },
-        published: false,
-        grading_type: "letter_grade",
-        notify_of_update: false,
-        lock_at: new Date().toISOString(),
-        // IMPORTANT: do NOT pass a time zone in the "due_at" field
-        due_at: `${examination.examDate}T23:59:59`,
-        // TODO: take the grading standard from TentaAPI
-        //       grading_standard_id: 1,
-      },
-    })
-    .then((r) => r.body);
-}
-
-/** Publish an assignment */
-async function publishAssignment(courseId, assignmentId) {
-  return canvas.requestUrl(
-    `courses/${courseId}/assignments/${assignmentId}`,
-    "PUT",
-    {
-      assignment: {
-        published: true,
-      },
-    }
-  );
-}
-
-async function unlockAssignment(courseId, assignmentId) {
-  const TOMORROW = new Date();
-  TOMORROW.setDate(TOMORROW.getDate() + 1);
-
-  return canvas.requestUrl(
-    `courses/${courseId}/assignments/${assignmentId}`,
-    "PUT",
-    {
-      assignment: {
-        lock_at: TOMORROW.toISOString(),
-        published: true,
-      },
-    }
-  );
-}
-
-async function lockAssignment(courseId, assignmentId) {
-  const YESTERDAY = new Date();
-  YESTERDAY.setDate(YESTERDAY.getDate() - 1);
-
-  return canvas.requestUrl(
-    `courses/${courseId}/assignments/${assignmentId}`,
-    "PUT",
-    {
-      assignment: {
-        lock_at: YESTERDAY.toISOString(),
-      },
-    }
-  );
-}
-
-// eslint-disable-next-line camelcase
-async function sendFile({ upload_url, upload_params }, filePath) {
-  const form = new FormData();
-
-  // eslint-disable-next-line camelcase
-  for (const key in upload_params) {
-    if (upload_params[key]) {
-      form.append(key, upload_params[key]);
-    }
-  }
-
-  form.append("attachment", fs.createReadStream(filePath));
-
-  return got.post({
-    url: upload_url,
-    body: form.stream,
-    headers: form.headers,
-    responseType: "json",
-  });
-}
-
-// TODO: Refactor this function and uploadExam to avoid requesting the endpoint
-//       "GET users/sis_user_id:${userId}" twice
-async function hasSubmission({ courseId, assignmentId, userId }) {
-  try {
-    const { body: user } = await canvas.get(`users/sis_user_id:${userId}`);
-    const { body: submission } = await canvas.get(
-      `courses/${courseId}/assignments/${assignmentId}/submissions/${user.id}`
-    );
-
-    return !submission.missing;
-  } catch (err) {
-    if (err.response?.statusCode === 404) {
-      return false;
-    }
-    throw err;
-  }
-}
-
-async function uploadExam(
-  filePath,
-  { courseId, assignmentId, userId, examDate }
-) {
-  try {
-    const { body: user } = await canvas.get(`users/sis_user_id:${userId}`);
-
-    // TODO: will return a 400 if the course is unpublished
-    const { body: slot } = await canvas.requestUrl(
-      `courses/${courseId}/assignments/${assignmentId}/submissions/${user.id}/files`,
-      "POST",
-      {
-        name: `${userId}.pdf`,
-      }
-    );
-
-    const { body: uploadedFile } = await sendFile(slot, filePath);
-
-    await canvas.requestUrl(
-      `courses/${courseId}/assignments/${assignmentId}/submissions/`,
-      "POST",
-      {
-        submission: {
-          submission_type: "online_upload",
-          user_id: user.id,
-          file_ids: [uploadedFile.id],
-          // IMPORTANT: do not pass the timezone in the "submitted_at" field
-          submitted_at: `${examDate}T08:00:00`,
-        },
-      }
-    );
-  } catch (err) {
-    if (err.response?.statusCode === 404) {
-      log.warn(`User ${userId} is missing in Canvas course ${courseId}`);
-    } else {
-      throw err;
-    }
-  }
-}
-
-async function getAuthorizationData(courseId, userId) {
-  const enrollments = await canvas
-    .list(`courses/${courseId}/enrollments`, { user_id: userId })
-    .toArray();
-
-  const roles = enrollments.map((enr) => enr.role_id);
-
-  const TEACHER = 4;
-  const EXAMINER = 10;
-
-  return {
-    authorized: roles.includes(TEACHER) || roles.includes(EXAMINER),
-    roles,
-  };
-}
-
-module.exports = {
-  getCourse,
-  publishCourse,
-  createHomepage,
-  getExaminationLadokId,
-  getValidAssignment,
-  createAssignment,
-  publishAssignment,
-  unlockAssignment,
-  lockAssignment,
-  hasSubmission,
-  uploadExam,
-  getAuthorizationData,
-};
